+++ ext/canvas/image_bitmap.rs
